fixed preivew
