+++ mochaccino-client/package.json
@@ -14,6 +14,7 @@
     "axios": "^0.27.2",
     "bootstrap": "^5.2.1",
     "react": "^18.2.0",
+    "react-json-pretty": "^2.2.0",
     "react-bootstrap": "^2.5.0",
     "react-dom": "^18.2.0",
     "react-icons": "^4.4.0",
